--- native/app/lib/context/index.ts
@@ -1,1 +1,0 @@
-export { Context as UiContext } from './ui';
